requirements.txt
Flask==3.1.0
boto3==1.24.0
python-dotenv==1.0.1
flask-cors==5.0.0
PyJWT